backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
pytz==2024.2
pyupbit==0.2.34
numpy==2.0.2
eventlet==0.36.1
